move trophy model to its own file
