+++ app/batid/models/__init__.py
@@ -4,3 +4,4 @@ from .feve import *  # noqa
 from .others import *  # noqa
 from .report import *  # noqa
 from .report_message import *  # noqa
+from .trophy import *  # noqa
